requirements.txt
# Core Streamlit
streamlit==1.29.0

# Firebase
firebase-admin==6.3.0

# Google Gemini AI
google-generativeai==0.3.1

# Environment & Configuration
python-dotenv==1.0.0

# Image Processing
Pillow==10.1.0

# Utilities
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1
isort==5.13.2

# Type Hints
types-requests==2.31.0.10

# Security
cryptography==41.0.7

# Data Validation
pydantic==2.5.3
email-validator==2.1.0

# Testing Helpers
faker==21.0.0
